requirements.txt
asgiref==3.6.0
autopep8==2.0.4
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-querycount==0.7.0
django-quill-editor==0.1.40
django-resized==1.0.2
django-silk==5.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gprof2dot==2022.7.29
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==23.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.11.1
PyJWT==1.7.1
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.6.0
zipp==3.11.0
channels
django-channels